Sort common files into packages
