+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.util.Ui;
+import duke.util.Storage;
+
+import duke.task.TaskList;
+
 import java.io.IOException;
 
 public abstract class Command {
