+++ src/main/java/duke/command/DoneCommand.java
@@ -1,13 +1,20 @@
+package duke.command;
+
+import duke.util.Ui;
+import duke.util.Storage;
+
+import duke.task.TaskList;
+
 import java.io.IOException;
 
-public class AddCommand extends Command {
-    public AddCommand(String input) {
+public class DoneCommand extends Command {
+    public DoneCommand(String input) {
         super(input);
     }
 
     @Override
     public TaskList execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
-        TaskList newTasks = tasks.addTask(input, ui);
+        TaskList newTasks = tasks.markTask(input, ui);
         storage.saveTasksToFile(newTasks);
         return newTasks;
     }
