+++ src/main/java/duke/util/Parser.java
@@ -1,3 +1,14 @@
+package duke.util;
+
+import duke.command.AddCommand;
+import duke.command.ByeCommand;
+import duke.command.Command;
+import duke.command.DeleteCommand;
+import duke.command.DoneCommand;
+import duke.command.DueCommand;
+import duke.command.ListCommand;
+import duke.command.OnDateCommand;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeParseException;
