+++ src/main/java/duke/util/Ui.java
@@ -1,3 +1,5 @@
+package duke.util;
+
 import java.util.Scanner;
 
 public class Ui {
@@ -20,7 +22,7 @@ public class Ui {
     public void showIntroduction() {
         System.out.print(LOGO);
         System.out.printf(FORMAT, LINE);
-        System.out.printf(FORMAT, "Hello there, I'm Duke!");
+        System.out.printf(FORMAT, "Hello there, I'm duke.Duke!");
         System.out.printf(FORMAT, "What can I do for you today?");
         System.out.printf(FORMAT, LINE);
     }
@@ -42,7 +44,7 @@ public class Ui {
     }
 
     public void showFileNotFoundError() {
-        System.out.printf(FORMAT, "This appears to be your first time using Duke.");
+        System.out.printf(FORMAT, "This appears to be your first time using duke.Duke.");
         System.out.printf(FORMAT, "A save file will be created to save your tasks when you first add a task.");
     }
